Modularize loader main
